extracted 2 more dtos
