+++ domain/src/main/java/be/xplore/notifyme/services/IVenueService.java
@@ -2,7 +2,6 @@ package be.xplore.notifyme.services;
 
 import be.xplore.notifyme.domain.User;
 import be.xplore.notifyme.domain.Venue;
-import be.xplore.notifyme.dto.GetVenueDto;
 import java.security.Principal;
 import java.util.List;
 
@@ -13,7 +12,7 @@ public interface IVenueService {
 
   Venue getVenue(long id);
 
-  List<GetVenueDto> getVenuesForUser(String userId);
+  List<Venue> getVenuesForUser(String userId);
 
   void makeUserVenueManager(String userId, Long venueId);
