+++ domain/src/main/java/be/xplore/notifyme/services/implementations/VenueService.java
@@ -3,13 +3,11 @@ package be.xplore.notifyme.services.implementations;
 import be.xplore.notifyme.domain.Address;
 import be.xplore.notifyme.domain.User;
 import be.xplore.notifyme.domain.Venue;
-import be.xplore.notifyme.dto.GetVenueDto;
 import be.xplore.notifyme.exception.CrudException;
 import be.xplore.notifyme.exception.SaveToDatabaseException;
 import be.xplore.notifyme.persistence.IVenueRepo;
 import be.xplore.notifyme.services.IVenueService;
 import java.security.Principal;
-import java.util.LinkedList;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -75,14 +73,8 @@ public class VenueService implements IVenueService {
    * @return list object containing all venues.
    */
   @Override
-  public List<GetVenueDto> getVenuesForUser(String userId) {
-    var venues = venueRepo.getAllByManagersIsContaining(userId);
-    List<GetVenueDto> venueDtos = new LinkedList<>();
-    for (Venue venue : venues) {
-      venueDtos.add(new GetVenueDto(venue.getId(), venue.getName(), venue.getDescription(),
-          venue.getAddress()));
-    }
-    return venueDtos;
+  public List<Venue> getVenuesForUser(String userId) {
+    return venueRepo.getAllByManagersIsContaining(userId);
   }
 
   /**
