+++ rest-api/src/main/java/be/xplore/notifyme/dto/line/GetLineDto.java
@@ -1,4 +1,4 @@
-package be.xplore.notifyme.dto;
+package be.xplore.notifyme.dto.line;
 
 import be.xplore.notifyme.domain.Line;
 import lombok.AllArgsConstructor;
